blockchain_framework/requirements.txt
flask
pycryptodome
gunicorn